requirements.txt
asgiref==3.5.0
black==22.3.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.4
Django==3.2.13
django-ckeditor==6.4.0
django-colorfield==0.7.0
django-contrib-comments==2.2.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-image-uploader-widget==0.1.6
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-taggit==3.0.0
filelock==3.7.0
flake8==4.0.1
identify==2.5.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
phonenumbers==8.12.48
Pillow==9.1.0
platformdirs==2.5.2
pre-commit==2.19.0
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
virtualenv==20.14.1
